+++ prototypes/DI_nav-ia-prototype_Figma-Make/src/app/imports/ResourceHub-2297-7065.tsx
@@ -1,9 +1,0 @@
-import { ResourceHubTile } from '../components/ResourceHubTile';
-
-export default function ResourceHub() {
-  return (
-    <div className="relative size-full" data-name="Resource Hub">
-      <ResourceHubTile />
-    </div>
-  );
-}